+++ catalog/app/containers/Bucket/PackageTree/PackageTree.tsx
@@ -575,6 +575,7 @@ function FileDisplayQuery({
   crumbs,
   ...props
 }: FileDisplayQueryProps) {
+  const { urls } = NamedRoutes.use()
   const fileQuery = GQL.useQuery(FILE_QUERY, { bucket, name, hash, path })
   return GQL.fold(fileQuery, {
     fetching: () => <FileDisplaySkeleton crumbs={crumbs} />,
@@ -582,6 +583,18 @@ function FileDisplayQuery({
       const file = d.package?.revision?.file
 
       if (!file) {
+        if (d.package?.revision?.dir) {
+          return (
+            <RRDom.Redirect
+              to={urls.bucketPackageTree(
+                bucket,
+                name,
+                props.hashOrTag,
+                s3paths.ensureSlash(path),
+              )}
+            />
+          )
+        }
         // eslint-disable-next-line no-console
         if (fileQuery.error) console.error(fileQuery.error)
         return (